Refactor API routes
