+++ routes/api_v1/assignments.js
@@ -3,7 +3,7 @@
  */
 
 const express = require('express');
-const db = require('../utils/db');
+const db = require('./../../utils/db');
 const router = express.Router();
 
 //TODO add echo support
