+++ utils/models.js
@@ -126,7 +126,7 @@ Users.hasOne(Teachers);
 UserLocal.belongsTo(Users);
 Users.hasOne(UserLocal);
 
-sequelize.sync();
+// sequelize.sync();
 
 module.exports = {
   Students,
